sermant-plugins/sermant-springboot-registry/springboot-registry-service/src/main/java/com/huawei/discovery/service/lb/rule/RandomLoadbalancer.java
/*
 * Copyright (C) 2022-2022 Huawei Technologies Co., Ltd. All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.huawei.discovery.service.lb.rule;

import com.huawei.discovery.entity.ServiceInstance;

import java.util.List;
import java.util.concurrent.ThreadLocalRandom;

/**
 * 随机负载均衡
 *
 * @author zhouss
 * @since 2022-09-28
 */
public class RandomLoadbalancer extends AbstractLoadbalancer {
    @Override
    protected ServiceInstance doChoose(String serviceName, List<ServiceInstance> instances) {
        final int index = ThreadLocalRandom.current().nextInt(instances.size());
        return instances.get(index);
    }

    @Override
    public String lbType() {
        return "Random";
    }
}
